requirements.txt
alembic==1.11.3
blinker==1.6.2
click==8.1.6
distlib==0.3.7
dnspython
email-validator==2.0.0.post2
filelock
Flask==2.2.5  
Flask-Cors==4.0.0
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
gunicorn
idna==3.4
importlib-metadata 
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow 
packaging==23.1
platformdirs==3.10.0
psycopg2-binary==2.9.7
python-dotenv 
SQLAlchemy==2.0.20
typing_extensions==4.7.1
virtualenv==20.24.2
Werkzeug 
WTForms==3.0.1
zipp